% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFAshiny.R
\name{EFAshiny}
\alias{EFAshiny}
\title{An user-friendly shiny application for exploratory factor analysis}
\usage{
EFAshiny(run = T, report_Apppath = F)
}
\arguments{
\item{run}{run the EFAshiny app or not. Default is true, and the APP will be lanuched.}

\item{report_Apppath}{report the directory of the APP in your system. If false, show the online APP link.}
}
\description{
The EFAshiny graphical user interface is designed to free users from scripting in R by wrapping together various packages for data management, factor analysis, and graphics.
}
\note{
EFAshiny can directly be used on-line. Check See Also section.
}
\examples{
if (interactive()) {
  EFAshiny::EFAshiny()
}
}
\seealso{
Online version: \url{https://psychilin.shinyapps.io/EFAshiny/}

Tutorial : \url{https://github.com/PsyChiLin/EFAshiny}
}
\author{
Chi-Lin Yu <psychilinyu@gmail.com>

Ching-Fan Sheu <csheu@mail.ncku.edu.tw>
}
