% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFAshiny.R
\name{EFAshiny}
\alias{EFAshiny}
\title{An user-friendly shiny application for exploratory factor analysis.}
\usage{
EFAshiny()
}
\description{
The EFAshiny graphical user interface in shiny (Chang, Cheng, Allaire, Xie, & McPherson, 2017) is designed to free users from scripting in R by wrapping together various packages, such as ggplot2 (Wickham, 2016) and psych (Revelle, 2017) R packages, for data management, factor analysis, and graphics. Easy-to-follow analysis flow and reasonable default settings avoiding common errors (Henson & Roberts, 2006) are provided. Results of analysis in tables and graphs are presented on-line and can be exported.
}
\examples{
if (interactive()) {
  EFAshiny()
}
}
\seealso{
\url{https://github.com/PsyChiLin/EFAshiny}
}
