% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFAshiny.R
\name{EFAshiny}
\alias{EFAshiny}
\title{EFAshiny is an user-friendly shiny application for exploratory factor analysis}
\usage{
EFAshiny()
}
\description{
EFAshiny is an user-friendly shiny application for exploratory factor analysis
}
\examples{
if (interactive()) {
  EFAshiny()
}
}
\seealso{
\url{https://github.com/PsyChiLin/EFAshiny}
}
