% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFAshiny.R
\name{EFAshiny}
\alias{EFAshiny}
\title{An user-friendly shiny application for exploratory factor analysis}
\usage{
EFAshiny()
}
\description{
The EFAshiny graphical user interface is designed to free users from scripting in R by wrapping together various packages for data management, factor analysis, and graphics.
}
\note{
EFAshiny can directly be used on-line. Check See Also section.
}
\examples{
if (interactive()) {
  EFAshiny()
}
}
\seealso{
Online version: \url{https://psychilin.shinyapps.io/EFAshiny/}

Tutorial : \url{https://github.com/PsyChiLin/EFAshiny}
}
\author{
Chi-Lin Yu <psychilinyu@gmail.com>

Ching-Fan Sheu <csheu@mail.ncku.edu.tw>
}
